feat: split landing and main page logic through router
