+++ frontend/src/components/layout/landing/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Landing';
