+++ frontend/src/components/layout/main/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Main';
